tools: grapher: refactor as a lib instead of main package
