+++ toolkit/tools/pkg/graph/grapher/config.go
@@ -1,0 +1,8 @@
+package grapher
+
+type Config struct {
+	Input            string
+	Output           string
+	StrictGoals      bool
+	StrictUnresolved bool
+}
